feat: add init settings
