+++ src/components/Header.jsx
@@ -4,6 +4,7 @@ import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
+import { Link } from 'react-router-dom';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -24,8 +25,12 @@ function Header() {
           <Typography variant="h5" className={classes.title}>
             Aruku App
           </Typography>
-          <Button color="inherit">Walk Settings</Button>
-          <Button color="inherit">Init Settings</Button>
+          <Button component={Link} to="/" color="inherit">
+            Walk Settings
+          </Button>
+          <Button component={Link} to="/Init" color="inherit">
+            Init Settings
+          </Button>
         </Toolbar>
       </AppBar>
     </div>
